More verbose error on duplicate file
I was stumped by the "file already exists" error message until I
realized it was a backup copy of my file leading to duplicate post
entries. This is a more verbose error message hinting towards that
as a potential error.

Ideally, we would have a possibility to exlude file patterns such as
*~ from being parsed.

Signed-off-by: Sebastian Spaeth <Sebastian@SSpaeth.de>