Allow to use customized atom plugin template
Previously we could customize the atom template name, but would always
search in the plugin directory (which is usually not
accessible/modifiable by the user in a regular install). Modify the
code, so that we search first in the custom 'templates' directory,
falling back to the builtin directory if the template name was not
found. This way the customized version gets preferred.

If no customized 'templates' dir exists, this will not cause a jinja2
error, the loader will simply ignore it and go directly to the builtin
version, so there is no need for explicit error checkin.

Signed-off-by: Sebastian Spaeth <Sebastian@SSpaeth.de>